+++ CHANGELOG.md
@@ -9,9 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - This changelog
 - Wish properties
+- Button to mark wish as fulfilled
 
-### Changed
-- Improved card design
+### Improved
+- Card design
 
 ### Fixed
  - Various small bugs